Initial implementation
The culmination of perhaps an hour spent modifying a copy of my PyPI
plugin (https://github.com/sopel-irc/sopel-pypi), which was the closest
thing I remembered writing to what this needed to do.

Not *super* extensively tested, especially on things like error cases
(mod not found or possible malformed JSON), nor have I stress-tested the
URL regex. A negative lookahead for `/` is all you get for now, because
oh look, it's 3am.